WIP BSP implementation. Going to switch stuff up
* added some code to collision map to try to fix partition overlaps
* added some code for debugging the collision map

I just realized that using the average position of the nodes causes the partition line to potentially change dramatically over time as new nodes are added/removed, which is currently resulting in hard-to-debug errors that are counter intuitive and creating a fragile sortof system. A quadtree or quadtree-like heuristic for partitioning would make errors easier to recreate and think about.